+++ interface/private_src/luat_audio_ec618.c
@@ -25,7 +25,7 @@ extern void audio_play_tts_default_fun(void *param);
 extern void audio_play_tts_set_resource_ex(void *address, void *sdk_id, void *read_resource_fun);
 extern void audio_play_global_init_ex(audio_play_event_cb_fun_t event_cb, audio_play_data_cb_fun_t data_cb, audio_play_default_fun_t play_file_fun, audio_play_default_fun_t play_tts_fun, void *user_param);
 extern void audio_play_global_init_with_task_priority(audio_play_event_cb_fun_t event_cb, audio_play_data_cb_fun_t data_cb, audio_play_default_fun_t play_file_fun, audio_play_default_fun_t play_tts_fun, void *user_param, uint8_t priority);
-extern int audio_play_write_blank_raw_ex(uint32_t multimedia_id, uint8_t cnt, uint8_t add_font);
+extern int audio_play_write_blank_raw_ex(uint8_t multimedia_id, uint8_t cnt, uint8_t add_font);
 void luat_audio_play_file_default_fun(void *param)
 {
 	audio_play_file_default_fun(param);
@@ -50,67 +50,67 @@ void luat_audio_play_global_init_with_task_priority(audio_play_event_cb_fun_t ev
 	audio_play_global_init_with_task_priority(event_cb, data_cb, play_file_fun, play_tts_fun, user_param, priority);
 }
 
-int luat_audio_play_multi_files(uint32_t multimedia_id, audio_play_info_t info[], uint32_t files_num)
+int luat_audio_play_multi_files(uint8_t multimedia_id, audio_play_info_t info[], uint32_t files_num)
 {
 	return audio_play_multi_files(multimedia_id, info, files_num);
 }
 
-uint8_t luat_audio_play_is_finish(uint32_t multimedia_id)
+uint8_t luat_audio_play_is_finish(uint8_t multimedia_id)
 {
 	return audio_play_is_finish(multimedia_id);
 }
 
-int luat_audio_play_stop(uint32_t multimedia_id)
+int luat_audio_play_stop(uint8_t multimedia_id)
 {
 	return audio_play_stop(multimedia_id);
 }
 
-int luat_audio_play_fast_stop(uint32_t multimedia_id)
+int luat_audio_play_fast_stop(uint8_t multimedia_id)
 {
 	return audio_play_fast_stop(multimedia_id);
 }
 
-int luat_audio_play_clear_stop_flag(uint32_t multimedia_id)
+int luat_audio_play_clear_stop_flag(uint8_t multimedia_id)
 {
 	return audio_play_clear_stop_flag(multimedia_id);
 }
 
-int luat_audio_play_pause_raw(uint32_t multimedia_id, uint8_t is_pause)
+int luat_audio_play_pause_raw(uint8_t multimedia_id, uint8_t is_pause)
 {
 	return audio_play_pause_raw(multimedia_id, is_pause);
 }
 
-int luat_audio_play_get_last_error(uint32_t multimedia_id)
+int luat_audio_play_get_last_error(uint8_t multimedia_id)
 {
 	return audio_play_get_last_error(multimedia_id);
 }
 
-int luat_audio_play_write_blank_raw(uint32_t multimedia_id, uint8_t cnt, uint8_t add_font)
+int luat_audio_play_write_blank_raw(uint8_t multimedia_id, uint8_t cnt, uint8_t add_font)
 {
 	return audio_play_write_blank_raw_ex(multimedia_id, cnt, add_font);
 }
 
-int luat_audio_play_start_raw(uint32_t multimedia_id, uint8_t audio_format, uint8_t num_channels, uint32_t sample_rate, uint8_t bits_per_sample, uint8_t is_signed)
+int luat_audio_play_start_raw(uint8_t multimedia_id, uint8_t audio_format, uint8_t num_channels, uint32_t sample_rate, uint8_t bits_per_sample, uint8_t is_signed)
 {
 	return audio_play_start_raw(multimedia_id, audio_format, num_channels, sample_rate, bits_per_sample, is_signed);
 }
 
-int luat_audio_play_write_raw(uint32_t multimedia_id, uint8_t *data, uint32_t len)
+int luat_audio_play_write_raw(uint8_t multimedia_id, uint8_t *data, uint32_t len)
 {
 	return audio_play_write_raw(multimedia_id, data, len);
 }
 
-int luat_audio_play_stop_raw(uint32_t multimedia_id)
+int luat_audio_play_stop_raw(uint8_t multimedia_id)
 {
 	return audio_play_stop_raw(multimedia_id);
 }
 
-int luat_audio_play_tts_text(uint32_t multimedia_id, void *text, uint32_t text_bytes)
+int luat_audio_play_tts_text(uint8_t multimedia_id, void *text, uint32_t text_bytes)
 {
 	return audio_play_tts_text(multimedia_id, text, text_bytes);
 }
 
-int luat_audio_play_tts_set_param(uint32_t multimedia_id, uint32_t param_id, uint32_t param_value)
+int luat_audio_play_tts_set_param(uint8_t multimedia_id, uint32_t param_id, uint32_t param_value)
 {
 	return audio_play_tts_set_param(multimedia_id, param_id, param_value);
 }
@@ -126,17 +126,17 @@ void luat_audio_play_set_bus_type(uint8_t bus_type)
 	audio_play_set_bus_type(bus_type);
 }
 
-void *luat_audio_play_get_stream(uint32_t multimedia_id)
+void *luat_audio_play_get_stream(uint8_t multimedia_id)
 {
 	return audio_play_get_stream(multimedia_id);
 }
 
-void luat_audio_play_set_user_lock(uint32_t multimedia_id, uint8_t onoff)
+void luat_audio_play_set_user_lock(uint8_t multimedia_id, uint8_t onoff)
 {
 	audio_play_set_user_lock(multimedia_id, onoff);
 }
 
-void luat_audio_play_debug_onoff(uint32_t multimedia_id, uint8_t onoff)
+void luat_audio_play_debug_onoff(uint8_t multimedia_id, uint8_t onoff)
 {
 	audio_play_debug_onoff(multimedia_id, onoff);
 }
